refactor: split restore point flow into stage per component fix: hook into TaskEnd instead of log completion
